fix: Add GameWorld SDK facade, telemetry, and GORNA tweaks
Introduce a high-level GameWorld facade in khora-sdk for safe ECS and asset access and update the Application API to use &mut GameWorld. Wire a telemetry Sender into RenderAgent to emit GpuReport events and add a builder helper (.with_telemetry_sender) and emit_telemetry() logic; add crossbeam-channel as a dependency and tests verifying telemetry wiring. Improve DCC (DccService) behavior by running an initial GORNA negotiation on first tick, introduce a SharedAgentList alias, add Drop to stop the service, and add an integration test ensuring initial negotiation applies budgets. Misc: add defaults and small simplifications to MetricStore/RingBuffer, expose additional lighting constants and struct fields in renderer/common, fix GPU monitor types, simplify map_err closures across render lanes, and update docs/examples to reflect these changes.

+++ crates/khora-infra/src/telemetry/gpu_monitor.rs
@@ -70,8 +70,8 @@ impl GpuMonitor {
             cpu_submission_time_us: Some(
                 (render_stats.cpu_render_submission_time_ms * 1000.0) as u32,
             ),
-            draw_calls: render_stats.draw_calls as u32,
-            triangles_rendered: render_stats.triangles_rendered as u32,
+            draw_calls: render_stats.draw_calls,
+            triangles_rendered: render_stats.triangles_rendered,
         };
 
         let mut last_stats = self.last_frame_stats.lock().unwrap();
